requirements.txt
websockets==3.2
discord.py==0.16.7
aiohttp==1.0.2
Pillow==4.1.1
